Restructured framework
- removed type mapping
- api now only handles byte[] as payload
- slf4j dependency removed
- old tests removed

+++ src/main/java/de/dk/ch/IDGenerator.java
@@ -17,5 +17,5 @@ public interface IDGenerator {
     *
     * @throws NoSuchElementException If no more id can be generated
     */
-   public long nextId() throws NoSuchElementException;
+   public byte nextId() throws NoSuchElementException;
 }
